fastAPI/requirements.txt
fastapi==0.89.1
mlflow==2.1.1
mlflow_skinny==1.30.0
numpy==1.22.3
pandas==1.4.4
pydantic==1.9.2
pymongo==4.3.3
python-dotenv==0.21.1
snowflake-connector-python==3.0.0
SQLAlchemy==1.4.41
uvicorn==0.16.0
boto3==1.26.28
ray==2.0.1
scikit_learn==1.2.1
xgboost==1.6.2